Move facebookPixel.ts from utils to services directory
